requirements.txt
googel-generativeai
deepgram-sdk
elevenlabs
